[JIT] Modify to_backend API so that it accepts wrapped modules
**Summary**
This commit modifies the `torch._C._jit_to_backend` function so that it
accepts `ScriptModules` as inputs. It already returns `ScriptModules`
(as opposed to C++ modules), so this makes sense and makes the API more
intuitive.

**Test Plan**
Continuous integration, which includes unit tests and out-of-tree tests
for custom backends.

**Fixes**
This commit fixes #41432.

ghstack-source-id: 21076f9d05950ca21e662daa8561f9931ba2f72f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41511